Use 'BigInt' type

+++ conf/db/migration/default/V4__add_simulation.sql
@@ -6,7 +6,7 @@ create table simulation
 (
     task_id          uuid                     not null,
     dashboard_id     uuid                     not null,
-    reached_modifier int                      not null,
+    reached_modifier bigint                   not null,
     created_at       timestamp with time zone not null,
     updated_at       timestamp with time zone
 );
